changes to all the test files for layout and nav
