+++ packages/terra-consumer-layout/tests/nightwatch/LayoutTestRoutes.jsx
@@ -7,14 +7,12 @@ import LayoutTests from './LayoutTests';
 // Test Cases
 import DexLayout from './DexLayout';
 import PortalLayout from './PortalLayout';
-import MemorialHerman from './memorial-herman/MemorialHerman';
 
 const routes = (
   <div>
     <Route path="/tests/layout-tests" component={LayoutTests} />
     <Route path="/tests/layout-tests/dex" component={DexLayout} />
     <Route path="/tests/layout-tests/portal" component={PortalLayout} />
-    <Route path="/tests/layout-tests/memlx" component={MemorialHerman} />
   </div>
 );
