--- packages/terra-consumer-nav/tests/nightwatch/Layout.scss
@@ -1,86 +1,0 @@
-@import '../../src/variables';
-
-:local {
-  body {
-    font-family: 'Roboto', sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    margin: 0;
-  }
-
-  .layout {
-    background: linear-gradient(#ddd, #999);
-    display: flex;
-    font-size: 14px;
-    height: 100vh;
-    left: 0;
-    max-width: 100vw;
-    overflow: hidden;
-    position: absolute;
-    top: 0;
-    width: 100vw;
-  }
-
-  .nav-container {
-    width: $terra-consumer--nav-container-width;
-  }
-
-  .main-container {
-    height: 100vh;
-    left: $terra-consumer--nav-container-width;
-    max-width: calc(100vw - #{$terra-consumer--nav-container-width});
-    overflow: scroll;
-    position: absolute;
-    text-align: left;
-    width: calc(100vw - #{$terra-consumer--nav-container-width});
-    @media screen and (max-width: $terra-medium-breakpoint) {
-      left: 20px;
-      max-width: 100vw;
-      width: 100vw;
-    }
-  }
-
-  .nav-burger {
-    background: none;
-    border: 0;
-    box-shadow: 0;
-    color: #000;
-    font-size: 20px;
-    left: 0;
-    padding: 0;
-    visibility: hidden;
-    @media screen and (max-width: $terra-medium-breakpoint) {
-      margin-right: 10px;
-      visibility: visible;
-    }
-
-    &:hover {
-      background: #eee;
-      cursor: pointer;
-    }
-  }
-
-  .footer {
-    padding-bottom: 30px;
-    padding-right: 40px;
-    padding-top: 20px;
-    text-align: right;
-    visibility: hidden;
-    width: inherit;
-
-    @media screen and (max-width: $terra-medium-breakpoint) {
-      visibility: visible;
-    }
-  }
-
-
-  .help-button-desktop {
-    bottom: 15px;
-    position: fixed;
-    right: 15px;
-
-    @media screen and (max-width: $terra-medium-breakpoint) {
-      display: none;
-    }
-  }
-}
