+++ packages/terra-consumer-nav/tests/nightwatch/NavTests.jsx
@@ -5,8 +5,8 @@ import React from 'react';
 const NavTests = () => (
   <div>
     <ul>
-      <li><a href="/#/tests/nav-tests/portal">PortalNav</a></li>
-      <li><a href="/#/tests/nav-tests/dex">DexNav</a></li>
+      <li><a href="/#/tests/nav-tests/simple-nav">Simple Nav</a></li>
+      <li><a href="/#/tests/nav-tests/help-button">Help Button</a></li>
     </ul>
   </div>
 );
